fix(lib): do the widening multiplications in the type that receives them
Three products were computed in a narrower type than the value they end up in:
the NTP floor in long where time_t is 64-bit, the hostname length in 16-bit int
on AVR, and RCSwitch's header pulse count in int where the counter is uint32_t.
None of them can overflow at the values in use; each now multiplies in the
receiving type, so the guarantee does not rest on the values staying small.

+++ lib/RCSwitch/src/RCSwitch.cpp
@@ -353,7 +353,7 @@ bool RCSwitch::receiveProtocol(const int32_t p, uint32_t changeCount) {
     BeginData = (pro.invertedSignal) ? (2) : (1);
     // Header pulse count correction for more than one
     if(pro.HeaderFactor > 1) {
-      BeginData += (pro.HeaderFactor - 1) * 2;
+      BeginData += (static_cast<uint32_t>(pro.HeaderFactor) - 1U) * 2U;
     }
   }
   // Assuming the longer pulse length is the pulse captured in timings[FirstTiming]
